+++ src/Sentry/Internal/Constants.cs
@@ -35,7 +35,7 @@ namespace Sentry.Internal
         // See: https://github.com/getsentry/sentry-release-registry
 #if ANDROID
         public const string SdkName = "sentry.dotnet.android";
-#elif IOS || MACCATALYST
+#elif __IOS__
         public const string SdkName = "sentry.dotnet.cocoa";
 #else
         public const string SdkName = "sentry.dotnet";